requirements.txt
labelme==5.1.1
matplotlib==3.5.3
mayavi==4.8.1
numpy==1.21.6
onnx==1.13.0
onnxsim==0.4.13
opencv_python==4.7.0.68
Pillow==9.4.0
pydicom==2.3.1
scikit_image==0.19.3
scipy==1.7.3
skimage==0.0
stl==0.0.3
thop==0.1.1.post2209072238
torch==1.7.0+cu110
torchsummary==1.5.1
tqdm==4.64.1
vtk==9.2.5
